Add files via upload
i have fixed "getPWM returns random numbers" issue

+++ Adafruit_PWMServoDriver.h
@@ -84,7 +84,7 @@ class Adafruit_PWMServoDriver {
   void setExtClk(uint8_t prescale);
   void setPWMFreq(float freq);
   void setOutputMode(bool totempole);
-  uint8_t getPWM(uint8_t num);
+  uint16_t getPWM(uint8_t num);
   void setPWM(uint8_t num, uint16_t on, uint16_t off);
   void setPin(uint8_t num, uint16_t val, bool invert=false);
   uint8_t readPrescale(void);
